requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
email-validator==2.1.0.post1
pycryptodome==3.20.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
PyJWT==2.9.0
qrcode[pil]==7.4.2
cryptography>=42.0.0
python-jose==3.3.0
pytest==8.3.3
